+++ system-contract-testing/utils/async.js
@@ -18,6 +18,4 @@ class Async {
   }
 }
 
-module.exports = {
-  Async
-};
+module.exports = Async;
