+++ system-contract-testing/utils/utils.js
@@ -993,6 +993,4 @@ class Utils {
   }
 }
 
-module.exports = {
-  Utils
-};
+module.exports = Utils;
